do delete

+++ JavaServer/src/main/java/ru/works/dont/touch/server/servicies/ImageService.java
@@ -10,10 +10,12 @@ import ru.works.dont.touch.server.exceptions.ExistsException;
 import ru.works.dont.touch.server.exceptions.NotExistsException;
 import ru.works.dont.touch.server.repositories.ImageRepository;
 
+import java.awt.image.BufferedImage;
 import java.io.*;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Stream;
 
 @Service
 public class ImageService {
@@ -40,9 +42,10 @@ public class ImageService {
     }
 
     @Transactional
-    public void deleteById(Long id) {
-        imageRepository.deleteById(id);
-        //TODO: добавить в remove возвращение изображения (если возможно) и удалять соответствующий файл
+    public Image deleteById(Long id) throws NotExistsException {
+        var img = findImageById(id);
+        getImageFile(img).delete();
+        return img;
     }
 
     /**
@@ -51,10 +54,19 @@ public class ImageService {
      * @param cardId the id of card
      */
     @Transactional
-    public void deleteByCardId(Long cardId) {
+    public Iterable<Image> deleteByCardId(Long cardId) throws NotExistsException {
+        if (!imageRepository.existsByCardId(cardId)){
+            throw new NotExistsException("Image with cardId not exist, id: "+cardId);
+        }
+        var imgs = imageRepository.findByCardId(cardId);
         imageRepository.deleteAllByCardId(cardId);
-        File dir = getDirectory(cardId);
-        dir.delete();
+        for (Image img : imgs) {
+            var fileImg = getImageFile(img);
+            if (fileImg.exists()){
+                fileImg.delete();
+            }
+        }
+        return imgs;
     }
 
 
@@ -118,4 +130,10 @@ public class ImageService {
         return new File(imageDirectory, cardId.toString());
     }
 
+
+    private File getImageFile(Image image){
+        return new File(imageDirectory +
+                "/CardId_" + image.getCardId()
+                + "/Image_" + image.getId());
+    }
 }
